pyproject.toml
[project]
name = "arca-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiocache>=0.12.3",
    "aiohttp>=3.13.3",
    "asgi-correlation-id>=4.3.4",
    "fastapi>=0.128.4",
    "fastmcp>=3.0.0b1",
    "google-genai>=1.62.0",
    "lancedb>=0.29.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-ulid>=3.1.0",
    "redis>=7.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
